starting cleaning the data for df1
I added code to clean the data for df1.  It creates df1_clean that deletes empty columns and changes the fips column to int64 from float64.  I also change the deaths column to int64 from float64.  This will help with integration later.  Fips are a unique code for the county.